+++ packages/flutter_firebase_ui/lib/login_view.dart
@@ -42,9 +42,7 @@ class _LoginViewState extends State<LoginView> {
   }
 
   _handleGoogleSignIn() async {
-    GoogleSignIn _googleSignIn = new GoogleSignIn();
-
-    GoogleSignInAccount googleUser = await _googleSignIn.signIn();
+    GoogleSignInAccount googleUser = await googleSignIn.signIn();
     if (googleUser != null) {
       GoogleSignInAuthentication googleAuth = await googleUser.authentication;
       if (googleAuth.accessToken != null) {
@@ -63,8 +61,8 @@ class _LoginViewState extends State<LoginView> {
   }
 
   _handleFacebookSignin() async {
-    var facebookLogin = new FacebookLogin();
-    FacebookLoginResult result = await facebookLogin.logInWithReadPermissions(['email']);
+    FacebookLoginResult result =
+        await facebookLogin.logInWithReadPermissions(['email']);
     if (result.accessToken != null) {
       try {
         FirebaseUser user = await _auth.signInWithFacebook(accessToken: result.accessToken.token);