+++ packages/flutter_firebase_ui/lib/utils.dart
@@ -1,11 +1,17 @@
 import 'dart:async';
 
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_facebook_login/flutter_facebook_login.dart';
+import 'package:google_sign_in/google_sign_in.dart';
 import 'package:meta/meta.dart';
 
 import 'package:flutter_firebase_ui/l10n/localization.dart';
 
 enum ProvidersTypes { email, google, facebook, twitter, phone }
+
+final GoogleSignIn googleSignIn = new GoogleSignIn();
+final FacebookLogin facebookLogin = new FacebookLogin();
 
 ProvidersTypes stringToProvidersType(String value) {
   if (value.toLowerCase().contains('facebook')) return ProvidersTypes.facebook;
@@ -133,4 +139,26 @@ Future<Null> showErrorDialog(BuildContext context, String message,
           ],
         ),
   );
+}
+
+Future<void> signOutProviders() async {
+  var currentUser = await FirebaseAuth.instance.currentUser();
+  if (currentUser != null) {
+    await signOut(currentUser.providerData);
+  }
+
+  return await FirebaseAuth.instance.signOut();
+}
+
+Future<dynamic> signOut(Iterable providers) async {
+  return Future.forEach(providers, (p) async {
+    switch (p.providerId) {
+      case 'facebook.com':
+        await facebookLogin.logOut();
+        break;
+      case 'google.com':
+        await googleSignIn.signOut();
+        break;
+    }
+  });
 }