+++ packages/video_vlc/ios/Classes/VideoPlayerController.h
@@ -1,0 +1,17 @@
+//
+//  VideoPlayerController.h
+//  Pods-Runner
+//
+//  Created by Yann-Cyril PELUD on 19/09/2018.
+//
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface VideoPlayerController : UIViewController<UIGestureRecognizerDelegate>
+@property(strong, nonatomic) NSString *urlString;
+
+@end
+
+NS_ASSUME_NONNULL_END