+++ packages/flutter_firebase_ui/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.12] - 27/09/2018
+* after trouble sign in pop
+
+## [0.0.11] - 27/09/2018
+* add function signOutProviders
+
+## [0.0.10] - 26/09/2018
+* fixed empty null on Providers
+* allow for usage with landscape orientation
+
 ## [0.0.9] - 07/09/2018
 * add mail verification option